refactor: ReservationRank record 변환
