+++ src/main/java/roomescape/dto/ReservationRank.java
@@ -1,0 +1,7 @@
+package roomescape.dto;
+
+import roomescape.domain.Reservation;
+
+public record ReservationRank(Reservation reservation, long rank) {
+
+}
